requirements.txt
numpy==1.21.0
pandas==1.3.0
protobuf==3.17.3
grpcio==1.38.1
grpcio-tools==1.38.1
PyYAML==5.4.1
torch==1.9.0
torchvision==0.10.0
cryptography==3.4.7
scikit-learn==0.24.2
pytest==6.2.4
flake8==3.9.2